tmux-smart-attach: Start server if not running
Currently, if the server is not already running,
list-sessions returns empty output
so we end up calling new-session.

When that starts up the server,
tmux-resurrect restores the saved sessions
causing new-session to fail because "main" is already used.

This changes tmux-smart-attach to start up the tmux session
if it's not already running in advance of all of this
so that the sessions have been restored by then.